requirements.txt
Pillow
pymongo
openpyxl